Reassemble C++ methods in the CppBridge constructor
This will break the edge-currency-monero project, but will provide a documented way to do what they are doing.
